+++ packages/db/index.d.ts
@@ -28,6 +28,12 @@ import type {
   ContentInput,
   ListPsaPinRequestOptions,
   ListPsaPinRequestResults,
+  UserStorageUsedInput,
+  UserStorageUsedOutput,
+  StorageUsedOutput,
+  UserTagInput,
+  EmailSentInput,
+  LogEmailSentInput,
 } from './db-client-types'
 
 export { gql }
@@ -38,7 +44,10 @@ export class DBClient {
   getMetricsValue (key: string): Promise<{ total: number }>
   upsertUser (user: UpsertUserInput): Promise<UpsertUserOutput>
   getUser (issuer: string): Promise<UserOutput>
-  getUsedStorage (userId: number): Promise<{ uploaded: number, pinned: number }>
+  getUserByEmail (email: string): Promise<UserOutput>
+  getStorageUsed (userId: number): Promise<StorageUsedOutput>
+  emailHasBeenSent(email: EmailSentInput): Promise<boolean>
+  logEmailSent(email : LogEmailSentInput): Promise<{id: string}>
   createUpload (data: CreateUploadInput): Promise<CreateUploadOutput>
   getUpload (cid: string, userId: number): Promise<UploadItemOutput>
   listUploads (userId: number, opts?: ListUploadsOptions): Promise<UploadItemOutput[]>
@@ -52,7 +61,7 @@ export class DBClient {
   getPinRequests ({ size }: { size: number }): Promise<Array<PinRequestItemOutput>>
   deletePinRequests (ids: Array<number>): Promise<void>
   createPinSyncRequests (pinSyncRequests: Array<string>): Promise<void>
-  getPinSyncRequests ({ to, after, size }: { to: string, after?: string, size?: number }): Promise<PinSyncRequestOutput>
+  getPinSyncRequests ({ to, after, size }: { to?: string, after?: string, size?: number }): Promise<PinSyncRequestOutput>
   deletePinSyncRequests (ids: Array<string>): Promise<void>
   getDeals (cid: string): Promise<Deal[]>
   getDealsForCids (cids: string[]): Promise<Record<string, Deal[]>>
@@ -66,5 +75,9 @@ export class DBClient {
   createContent (content: ContentInput, opt?: {updatePinRequests?: boolean}) : Promise<string>
   deleteKey (id: number): Promise<void>
   query<T, V>(document: RequestDocument, variables: V): Promise<T>
+  createUserTag(userId: number, tag: UserTagInput): Promise<boolean>
   getUserTags (userId: number): Promise<{ tag: string, value: string }[]>
 }
+
+export function parseTextToNumber(n: string): number
+export { EMAIL_TYPE } from './constants.js'